requirements.txt
opencv-python
pillow
numpy
pyyaml
easyocr